move checkPartitionTransformCompatibleWithSpec to IcebergPartitionUtils
